added linkedin article
