+++ src/Components/Home/ProjectsSummary.tsx
@@ -1,18 +1,18 @@
-import ProjectsItem from "./ProjectsItem";
+import LinkItem from "./LinkItem";
 
 function ProjectsSummary() {
   return (
     <div className="mt-8">
       <h3 className="text-primary text-xl">Projects</h3>
-      <ProjectsItem
+      <LinkItem
         title="Cloud Resume Challenge"
         link="https://github.com/Lesmol/CloudResumeChallenge_AWS"
       />
-      <ProjectsItem
+      <LinkItem
         title="Varsity Registration"
         link="https://github.com/Lesmol/varsity_registration"
       />
-      <ProjectsItem
+      <LinkItem
         title="Techspo Website"
         link="https://github.com/Lesmol/Techspo_Website"
       />
