+++ src/UI/Home.tsx
@@ -5,6 +5,7 @@ import UserSummary from "../Components/Home/UserSummary.jsx";
 import About from "../Components/Home/About.jsx";
 import EmailMe from "../Components/Home/EmailMe.jsx";
 import ProjectsSummary from "../Components/Home/ProjectsSummary.jsx";
+import Articles from "../Components/Home/Articles.js";
 
 function Home() {
   return (
@@ -37,6 +38,13 @@ function Home() {
         transition={{ duration: 0.6, delay: 0.3 }}
       >
         <ProjectsSummary />
+      </motion.div>
+      <motion.div
+        initial={{ y: 20, opacity: 0 }}
+        animate={{ y: 0, opacity: 1 }}
+        transition={{ duration: 0.6, delay: 0.4 }}
+      >
+        <Articles />
       </motion.div>
     </div>
   );
